requirements.txt
azure-common==1.1.28
azure-core==1.31.0
azure-identity==1.18.0
azure-mgmt-core==1.4.0
azure-mgmt-loganalytics==12.0.0
azure-monitor-query==1.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
DateTime==5.5
idna==3.10
isodate==0.6.1
msal==1.31.0
msal-extensions==1.2.0
msrest==0.7.1
numpy==2.1.1
oauthlib==3.2.2
pandas==2.2.3
portalocker==2.10.1
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
zope.interface==7.0.3
